Moved maintenance tasks to Huey worker
