+++ .gitignore
@@ -2,6 +2,7 @@ __pycache__
 .DS_Store
 .venv
 .pytest_cache
+.ruff_cache
 *.pyc
 *.pyo
 .vscode
