+++ redash/tasks/queries/__init__.py
@@ -1,8 +1,1 @@
-from .maintenance import (
-    refresh_queries,
-    refresh_schemas,
-    cleanup_query_results,
-    empty_schedules,
-    remove_ghost_locks,
-)
 from .execution import execute_query, enqueue_query
